+++ packages/angular_devkit/core/src/json/schema/registry.ts
@@ -13,7 +13,7 @@ import * as https from 'https';
 import { Observable, from, isObservable } from 'rxjs';
 import { map } from 'rxjs/operators';
 import * as Url from 'url';
-import { BaseException } from '../../exception/exception';
+import { BaseException } from '../../exception';
 import { PartiallyOrderedSet, deepCopy } from '../../utils';
 import { JsonArray, JsonObject, JsonValue, isJsonObject } from '../utils';
 import {
